Clean some of the many new warnings from clang 10

+++ 1D/Bernoulli_IT.cpp
@@ -19,15 +19,15 @@ auto run(size_t N, size_t T, double theta, bool v) -> double {
     for (size_t i = 0; i < N; ++i) etas.push_back(prng.uniform_int(funs.size()));
 
     vector<double> fs(T);
-    tbb::parallel_for(size_t(0), T, [&](size_t i) { fs[i] = indexed_compose(funs, etas, double(i) / T); });
+    tbb::parallel_for(size_t(0), T, [&](size_t i) { fs[i] = indexed_compose(funs, etas, double(i) / double(T)); });
 
     if (v)
-        for (size_t i = 0; i < T; ++i) cout << double(i) / T << ' ' << fs[i] << '\n';
+        for (size_t i = 0; i < T; ++i) cout << double(i) / double(T) << ' ' << fs[i] << '\n';
 
     double s1 = 0.0, s2 = 0.0;
     for (auto u : fs) s1 += cos(2 * M_PI * u);
     for (auto u : fs) s2 += cos(2 * M_PI * u) * cos(2 * M_PI * u);
-    return s2 / T - s1 * s1 / double(T * T);
+    return s2 / double(T) - s1 * s1 / double(T * T);
 }
 
 auto main(int argc, char **argv) -> int {
